# src/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pathofmodifiers",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/api_v1/item/{itemId}": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Get Item",
        "description": "Get item by key and value for \"itemId\".\n\nAlways returns one item.",
        "operationId": "get_item",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Itemid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "items"
        ],
        "summary": "Update Item",
        "description": "Update an item by key and value for \"itemId\".\n\nReturns the updated item.",
        "operationId": "update_item",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Itemid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "items"
        ],
        "summary": "Delete Item",
        "description": "Delete an item by key and value for \"itemId\".\n\nReturns a message indicating the item was deleted.\nAlways deletes one item.",
        "operationId": "delete_item",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Itemid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Items-Delete Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api_v1/item/latest_item_id/": {
      "get": {
        "tags": [
          "items",
          "latest_item_id"
        ],
        "summary": "Get Latest Item Id",
        "description": "Get the latest \"itemId\"\n\nCan only be used safely on an empty table or directly after an insertion.",
        "operationId": "get_latest_item_id",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Items-Get Latest Item Id"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/api_v1/item/": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Get All Items",
        "description": "Get all items.\n\nReturns a list of all items.",
        "operationId": "get_all_items",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Item"
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Item"
                      }
                    }
                  ],
                  "title": "Response Items-Get All Items"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "items"
        ],
        "summary": "Create Item",
        "description": "Create one or a list of new items.\n\nReturns the created item or list of items.",
        "operationId": "create_item",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "return_nothing",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return Nothing"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ItemCreate"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ItemCreate"
                    }
                  }
                ],
                "title": "Item"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ItemCreate"
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ItemCreate"
                      }
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Items-Create Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Influences": {
        "properties": {
          "elder": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elder"
          },
          "shaper": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shaper"
          },
          "crusader": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crusader"
          },
          "redeemer": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redeemer"
          },
          "hunter": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hunter"
          },
          "warlord": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warlord"
          }
        },
        "type": "object",
        "title": "Influences"
      },
      "Item": {
        "properties": {
          "stashId": {
            "type": "string",
            "title": "Stashid"
          },
          "gameItemId": {
            "type": "string",
            "title": "Gameitemid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "iconUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iconurl"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "typeLine": {
            "type": "string",
            "title": "Typeline"
          },
          "baseType": {
            "type": "string",
            "title": "Basetype"
          },
          "ilvl": {
            "type": "integer",
            "title": "Ilvl"
          },
          "rarity": {
            "type": "string",
            "title": "Rarity"
          },
          "identified": {
            "type": "boolean",
            "title": "Identified",
            "default": true
          },
          "forumNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forumnote"
          },
          "currencyAmount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currencyamount"
          },
          "currencyId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currencyid"
          },
          "corrupted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrupted"
          },
          "delve": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delve"
          },
          "fractured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fractured"
          },
          "synthesized": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthesized"
          },
          "replica": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replica"
          },
          "elder": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elder"
          },
          "shaper": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shaper"
          },
          "influences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Influences"
              },
              {
                "type": "null"
              }
            ]
          },
          "searing": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searing"
          },
          "tangled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tangled"
          },
          "isRelic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isrelic"
          },
          "prefixes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefixes"
          },
          "suffixes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suffixes"
          },
          "foilVariation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foilvariation"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "itemId": {
            "type": "integer",
            "title": "Itemid"
          }
        },
        "type": "object",
        "required": [
          "stashId",
          "gameItemId",
          "league",
          "typeLine",
          "baseType",
          "ilvl",
          "rarity",
          "createdAt",
          "itemId"
        ],
        "title": "Item"
      },
...
      "ItemCreate": {
        "properties": {
          "stashId": {
            "type": "string",
            "title": "Stashid"
          },
          "gameItemId": {
            "type": "string",
            "title": "Gameitemid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "iconUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iconurl"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "typeLine": {
            "type": "string",
            "title": "Typeline"
          },
          "baseType": {
            "type": "string",
            "title": "Basetype"
          },
          "ilvl": {
            "type": "integer",
            "title": "Ilvl"
          },
          "rarity": {
            "type": "string",
            "title": "Rarity"
          },
          "identified": {
            "type": "boolean",
            "title": "Identified",
            "default": true
          },
          "forumNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forumnote"
          },
          "currencyAmount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currencyamount"
          },
          "currencyId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currencyid"
          },
          "corrupted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrupted"
          },
          "delve": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delve"
          },
          "fractured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fractured"
          },
          "synthesized": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthesized"
          },
          "replica": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replica"
          },
          "elder": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elder"
          },
          "shaper": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shaper"
          },
          "influences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Influences"
              },
              {
                "type": "null"
              }
            ]
          },
          "searing": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searing"
          },
          "tangled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tangled"
          },
          "isRelic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isrelic"
          },
          "prefixes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefixes"
          },
          "suffixes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suffixes"
          },
          "foilVariation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foilvariation"
          }
        },
        "type": "object",
        "required": [
          "stashId",
          "gameItemId",
          "league",
          "typeLine",
          "baseType",
          "ilvl",
          "rarity"
        ],
        "title": "ItemCreate"
      },
...
      "ItemUpdate": {
        "properties": {
          "stashId": {
            "type": "string",
            "title": "Stashid"
          },
          "gameItemId": {
            "type": "string",
            "title": "Gameitemid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "iconUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iconurl"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "typeLine": {
            "type": "string",
            "title": "Typeline"
          },
          "baseType": {
            "type": "string",
            "title": "Basetype"
          },
          "ilvl": {
            "type": "integer",
            "title": "Ilvl"
          },
          "rarity": {
            "type": "string",
            "title": "Rarity"
          },
          "identified": {
            "type": "boolean",
            "title": "Identified",
            "default": true
          },
          "forumNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forumnote"
          },
          "currencyAmount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currencyamount"
          },
          "currencyId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currencyid"
          },
          "corrupted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrupted"
          },
          "delve": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delve"
          },
          "fractured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fractured"
          },
          "synthesized": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthesized"
          },
          "replica": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replica"
          },
          "elder": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elder"
          },
          "shaper": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shaper"
          },
          "influences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Influences"
              },
              {
                "type": "null"
              }
            ]
          },
          "searing": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searing"
          },
          "tangled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tangled"
          },
          "isRelic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isrelic"
          },
          "prefixes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefixes"
          },
          "suffixes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suffixes"
          },
          "foilVariation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foilvariation"
          }
        },
        "type": "object",
        "required": [
          "stashId",
          "gameItemId",
          "league",
          "typeLine",
          "baseType",
          "ilvl",
          "rarity"
        ],
        "title": "ItemUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/api_v1/login/access-token"
          }
        }
      }
    }
  }
}